Made the argument errors a bit more ergonomic to work with

+++ src/compile/error.rs
@@ -5,8 +5,8 @@ use crate::sxp;
 pub enum Error {
   DottedListError,
   CannotCall(String),
-  TooFewArgs(String, i32),
-  TooManyArgs(String, i32),
+  TooFewArgs(String, usize),
+  TooManyArgs(String, usize),
 }
 
 impl From<sxp::dotted::TryFromDottedExprError> for Error {
